feat: divided server and client actions
