+++ src/app/(authorized)/search/SearchContent/actions.ts
@@ -1,42 +1,21 @@
-import { cookies } from 'next/headers';
 import { getUsersOptions } from '@/app/lib/actionOptions';
+import { accessToken } from '@/utils/serverActions';
+import {
+  GetFollowersDataType,
+  UserFollowingDataType,
+} from '@/app/(not_authorized)/(public-info)/public-post-page/[id]/types';
 
 const baseUrl = process.env.NEXT_PUBLIC_BASE_URL;
 
-export const getUsers = async (
-  search: string,
-  pageSize: number = 12,
-  pageNumber: number = 1,
-  cursor: number = 0
-) => {
-  const accessToken = cookies().get('accessToken')?.value;
-  const apiUrl =
-    baseUrl +
-    `users?search=${search}&pageSize=${pageSize}&pageNumber=${pageNumber}&cursor=${cursor}`;
-  try {
-    const response = await fetch(apiUrl, getUsersOptions(accessToken));
-    if (!response.ok) {
-      console.error('Error:', response.statusText);
-      return null;
-    }
-    const data = await response.json();
-    return data;
-  } catch (error) {
-    console.error('Error fetching data:', error);
-    return null;
-  }
-};
-
 export const getUserInfo = async (userName: string) => {
-  const accessToken = cookies().get('accessToken')?.value;
   const apiUrl = baseUrl + `users/${userName}`;
   try {
-    const response = await fetch(apiUrl, getUsersOptions(accessToken));
+    const response = await fetch(apiUrl, getUsersOptions(accessToken()));
     if (!response.ok) {
       console.error('Error:', response.statusText);
       return null;
     }
-    const data = await response.json();
+    const data: UserFollowingDataType = await response.json();
     return data;
   } catch (error) {
     console.error('Error fetching data:', error);
@@ -51,17 +30,16 @@ export const getUserFollowers = async (
   pageNumber: number = 1,
   cursor: number = 0
 ) => {
-  const accessToken = cookies().get('accessToken')?.value;
   const apiUrl =
     baseUrl +
     `users/${userName}/followers?search=${search}&pageSize=${pageSize}&pageNumber=${pageNumber}&cursor=${cursor}`;
   try {
-    const response = await fetch(apiUrl, getUsersOptions(accessToken));
+    const response = await fetch(apiUrl, getUsersOptions(accessToken()));
     if (!response.ok) {
       console.error('Error:', response.statusText);
       return null;
     }
-    const data = await response.json();
+    const data: GetFollowersDataType = await response.json();
     return data;
   } catch (error) {
     console.error('Error fetching data:', error);
@@ -76,17 +54,16 @@ export const getUserFollowing = async (
   pageNumber: number = 1,
   cursor: number = 0
 ) => {
-  const accessToken = cookies().get('accessToken')?.value;
   const apiUrl =
     baseUrl +
     `users/${userName}/following?search=${search}&pageSize=${pageSize}&pageNumber=${pageNumber}&cursor=${cursor}`;
   try {
-    const response = await fetch(apiUrl, getUsersOptions(accessToken));
+    const response = await fetch(apiUrl, getUsersOptions(accessToken()));
     if (!response.ok) {
       console.error('Error:', response.statusText);
       return null;
     }
-    const data = await response.json();
+    const data: GetFollowersDataType = await response.json();
     return data;
   } catch (error) {
     console.error('Error fetching data:', error);
